Extract promotion bridge panel
